Bug 2027694 - Remove manual DomainMatchingMode cast. r=dom-storage-reviewers,jari
Differential Revision: https://phabricator.services.mozilla.com/D292859

UltraBlame original commit: 44056e913a48a481eb85a699d48059769dfd1a17

+++ dom/storage/SessionStorageManager.cpp
@@ -2,8 +2,6 @@
 
 
 
-
-
 #include "SessionStorageManager.h"
 
 #include "SessionStorage.h"
@@ -655,8 +653,7 @@ void SessionStorageManager::ClearStorages(
       return;
     }
 
-    mActor->SendClearStorages(aPattern, nsCString(aOriginScope),
-                              static_cast<uint32_t>(aMode));
+    mActor->SendClearStorages(aPattern, nsCString(aOriginScope), aMode);
   }
 
   ClearStoragesInternal(aPattern, aOriginScope, aMode);
